[now dev] Move helper modules to `src/util/dev`
This organization is consistent with how other commands are structured.
