+++ src/util/dev/static-builder.ts
